enrollment-notes: persist storeDate with T format
