--- core/src/main/java/org/hisp/dhis/android/core/data/database/GenericDbDateColumnAdapter.java
@@ -1,74 +1,0 @@
-/*
- * Copyright (c) 2017, University of Oslo
- *
- * All rights reserved.
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- * Redistributions of source code must retain the above copyright notice, this
- * list of conditions and the following disclaimer.
- *
- * Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following disclaimer in the documentation
- * and/or other materials provided with the distribution.
- * Neither the name of the HISP project nor the names of its contributors may
- * be used to endorse or promote products derived from this software without
- * specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package org.hisp.dhis.android.core.data.database;
-
-import android.content.ContentValues;
-import android.database.Cursor;
-
-import com.gabrielittner.auto.value.cursor.ColumnTypeAdapter;
-
-import org.hisp.dhis.android.core.common.SafeDateFormat;
-
-import java.text.ParseException;
-import java.util.Date;
-
-public abstract class GenericDbDateColumnAdapter implements ColumnTypeAdapter<Date> {
-
-    private final SafeDateFormat dateFormat;
-
-    GenericDbDateColumnAdapter(SafeDateFormat dateFormat) {
-        this.dateFormat = dateFormat;
-    }
-
-    @Override
-    public Date fromCursor(Cursor cursor, String columnName) {
-        // infer index from column name
-        int columnIndex = cursor.getColumnIndex(columnName);
-        String sourceDate = cursor.getString(columnIndex);
-
-        Date date = null;
-        if (sourceDate != null) {
-            try {
-                date = dateFormat.parse(sourceDate);
-            } catch (ParseException parseException) {
-                // wrap checked exception into unchecked
-                throw new RuntimeException(parseException);
-            }
-        }
-
-        return date;
-    }
-
-    @Override
-    public void toContentValues(ContentValues contentValues, String columnName, Date date) {
-        if (date != null) {
-            contentValues.put(columnName, dateFormat.format(date));
-        }
-    }
-}
